só comentei tudo relacionado a DataHelper
Vou tentar usar o Room
