+++ .idea/.name
@@ -1,1 +1,1 @@
-Filmapp
+filmapp
