--- README.md
@@ -1,14 +1,0 @@
-# Filmapp
-## O Filmapp é um aplicativo de catalogação de filmes organizados em listas semelhantes ao Letterboxd.
-
-## Neste App, o usuário deve ser capaz de executar algumas funcionalidades básicas como:
-
-### 1-Adicionar filmes no catálogo. 
-
-### 2-Listar quais os filmes que já foram adicionados no catálogo. (Ele sempre deve ser atualizado caso haja alguma nova alteração).
-
-### 3-Excluir filmes do catálogo.
-
-### 4-Alterar informações dos filmes que já foram adicionados.
-
-### 5-Pesquisar filmes com a ferramenta de busca tanto para procurar filmes a serem adicionados tanto para pesquisar determinado filme dentro do seu catálogo.
