+++ app/src/main/AndroidManifest.xml
@@ -12,6 +12,9 @@
         android:theme="@style/Theme.Filmapp"
         tools:targetApi="31">
         <activity
+            android:name=".RegisterActivity"
+            android:exported="false" />
+        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
