+++ settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "Filmapp"
+rootProject.name = "filmapp"
 include ':app'
