--- src/main/java/com/example/filmapp/MainActivity.java
@@ -1,46 +1,0 @@
-package com.example.filmapp;
-
-import androidx.appcompat.app.AppCompatActivity;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.Toast;
-import com.example.filmapp.R;
-
-public class MainActivity extends AppCompatActivity {
-
-    private EditText editTextUsername;
-    private EditText editTextPassword;
-    private Button buttonLogin;
-
-    // Banco de dados ainda não implementado
-    private DatabaseHelper databaseHelper;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        editTextUsername = findViewById(R.id.editTextUsername);
-        editTextPassword = findViewById(R.id.editTextPassword);
-        buttonLogin = findViewById(R.id.buttonLogin);
-
-        // Inicializar o banco de dados
-        databaseHelper = new DatabaseHelper(this);
-
-        buttonLogin.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                String inputUsername = editTextUsername.getText().toString();
-                String inputPassword = editTextPassword.getText().toString();
-
-                if (databaseHelper.isValidUser(inputUsername, inputPassword)) {
-                    Toast.makeText(MainActivity.this, "Login successful", Toast.LENGTH_SHORT).show();
-                } else {
-                    Toast.makeText(MainActivity.this, "Login failed", Toast.LENGTH_SHORT).show();
-                }
-            }
-        });
-    }
-}
